refactored code to move chat functions in different file
